module_27_postgres_migrations/hw/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
psycopg2-binary==2.9.9
gunicorn==21.2.0
requests==2.31.0